+++ src/KeyVault/KeyVault/Commands/NewAzureManagedHsm.cs
@@ -1,7 +1,22 @@
-﻿using Microsoft.Azure.Commands.KeyVault.Helpers;
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
+
 using Microsoft.Azure.Commands.KeyVault.Models;
 using Microsoft.Azure.Commands.KeyVault.Properties;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
+using Microsoft.WindowsAzure.Commands.Utilities.Common;
+
 using System;
 using System.Collections;
 using System.Management.Automation;
@@ -47,7 +62,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Commands
         [Parameter(Mandatory = true,
             Position = 2,
             ValueFromPipelineByPropertyName = true,
-            HelpMessage = "Specifies the Azure region in which to create the key vault. Use the command Get-AzResourceProvider with the ProviderNamespace parameter to see your choices.")]
+            HelpMessage = "Specifies the Azure region in which to create the managed HSM pool. Use the command Get-AzResourceProvider with the ProviderNamespace parameter to see your choices.")]
         [LocationCompleter("Microsoft.KeyVault/managedHSMs")]
         [ValidateNotNullOrEmpty()]
         public string Location { get; set; }
@@ -63,6 +78,16 @@ namespace Microsoft.Azure.Commands.KeyVault.Commands
             HelpMessage = "Specifies the SKU of the managed HSM instance.")]
         [PSArgumentCompleter("StandardB1", "CustomB32")]
         public string Sku { get; set; }
+
+        [Parameter(Mandatory = false,
+            HelpMessage = "Specifies how long the deleted managed hsm pool is retained, and how long until the managed hsm pool in the deleted state can be purged. The default is " + Constants.DefaultSoftDeleteRetentionDaysString + " days.")]
+        [ValidateRange(Constants.MinSoftDeleteRetentionDays, Constants.MaxSoftDeleteRetentionDays)]
+        [ValidateNotNullOrEmpty]
+        public int SoftDeleteRetentionInDays { get; set; }
+
+        [Parameter(Mandatory = false,
+            HelpMessage = "specifying whether protection against purge is enabled for this managed HSM pool. The setting is effective only if soft delete is also enabled. Enabling this functionality is irreversible.")]
+        public SwitchParameter EnablePurgeProtection { get; set; }
 
         [Parameter(Mandatory = false,
             ValueFromPipelineByPropertyName = true,
@@ -93,7 +118,13 @@ namespace Microsoft.Azure.Commands.KeyVault.Commands
                     TenantId = GetTenantId(),
                     Tags = this.Tag,
                     Administrator = this.Administrator,
-                    SkuFamilyName = DefaultManagedHsmSkuFamily
+                    SkuFamilyName = DefaultManagedHsmSkuFamily,
+                    // If retention days is not specified, use the default value
+                    SoftDeleteRetentionInDays = this.IsParameterBound(c => c.SoftDeleteRetentionInDays)
+                            ? SoftDeleteRetentionInDays
+                            : Constants.DefaultSoftDeleteRetentionDays,
+                    // false is not accepted
+                    EnablePurgeProtection = this.EnablePurgeProtection.IsPresent ? true : (bool?)null, 
                 };
 
                 this.WriteObject(KeyVaultManagementClient.CreateNewManagedHsm(vaultCreationParameter, GraphClient));