+++ src/KeyVault/KeyVault/Models/VaultManagementClient.cs
@@ -431,7 +431,6 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
                 }
                 properties.TenantId = parameters.TenantId;
                 properties.InitialAdminObjectIds = parameters.Administrator;
-                properties.HsmUri = "";
                 properties.EnableSoftDelete = parameters.EnableSoftDelete;
                 properties.SoftDeleteRetentionInDays = parameters.SoftDeleteRetentionInDays;
                 properties.EnablePurgeProtection = parameters.EnablePurgeProtection;
@@ -456,7 +455,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         }
 
         /// <summary>
-        /// Get an existing Managed HSM. Returns null if vault is not found.
+        /// Get an existing managed HSM. Returns null if managed HSM is not found.
         /// </summary>
         /// <param name="managedHsmName">managed HSM name</param>
         /// <param name="resourceGroupName">resource group name</param>
@@ -475,18 +474,34 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
 
                 return new PSManagedHsm(response, graphClient);
             }
-            catch (ManagedHsmErrorException ce)
+            catch (ManagedHsmErrorException ce) when (ce.IsNotFoundException())
             {
-                if (ce.Response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return null;
-                }
-                throw;
+                return null;
             }
         }
 
+        public PSDeletedManagedHsm GetDeletedManagedHsm(string managedHsmName, string location)
+        {
+            if (string.IsNullOrWhiteSpace(managedHsmName))
+                throw new ArgumentNullException("HsmName");
+            if (string.IsNullOrWhiteSpace(location))
+                throw new ArgumentNullException("Location");
+
+            try
+            {
+                var response = KeyVaultManagementClient.ManagedHsms.GetDeleted(managedHsmName, location);
+
+                return new PSDeletedManagedHsm(response);
+            }
+            catch (ManagedHsmErrorException ce) when (ce.IsNotFoundException())
+            {
+                return null;
+            }
+        }
+
+
         /// <summary>
-        /// List all existing Managed HSMs. Returns null if vault is not found.
+        /// List all existing Managed HSMs.
         /// </summary>
         /// <param name="resourceGroupName">resource group name</param>
         /// <param name="graphClient">the active directory client</param>
@@ -499,7 +514,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
 
         private List<PSManagedHsm> ListManagedHsmsByResourceGroup(string resourceGroupName, IMicrosoftGraphClient graphClient = null)
         {
-            List<PSManagedHsm> managedHsms = new List<PSManagedHsm>(); ;
+            List<PSManagedHsm> managedHsms = new List<PSManagedHsm>();
             IPage<ManagedHsm> response = KeyVaultManagementClient.ManagedHsms.ListByResourceGroupAsync(resourceGroupName).GetAwaiter().GetResult();
             foreach (var managedHsm in response)
             {
@@ -521,7 +536,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
 
         private List<PSManagedHsm> ListManagedHsmsBySubscription(IMicrosoftGraphClient graphClient = null)
         {
-            List<PSManagedHsm> managedHsms = new List<PSManagedHsm>(); ;
+            List<PSManagedHsm> managedHsms = new List<PSManagedHsm>();
             IPage<ManagedHsm> response = KeyVaultManagementClient.ManagedHsms.ListBySubscriptionAsync().GetAwaiter().GetResult();
 
             foreach (var managedHsm in response)
@@ -542,8 +557,31 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             return managedHsms;
         }
 
+        public List<PSDeletedManagedHsm> ListDeletedManagedHsms()
+        {
+            List<PSDeletedManagedHsm> deletedManagedHsms = new List<PSDeletedManagedHsm>();
+            IPage<DeletedManagedHsm> response = KeyVaultManagementClient.ManagedHsms.ListDeleted();
+
+            foreach (var deletedManagedHsm in response)
+            {
+                deletedManagedHsms.Add(new PSDeletedManagedHsm(deletedManagedHsm));
+            }
+
+            while (response?.NextPageLink != null)
+            {
+                response = KeyVaultManagementClient.ManagedHsms.ListDeletedNext(response.NextPageLink);
+
+                foreach (var deletedManagedHsm in response)
+                {
+                    deletedManagedHsms.Add(new PSDeletedManagedHsm(deletedManagedHsm));
+                }
+            }
+
+            return deletedManagedHsms;
+        }
+       
         /// <summary>
-        /// Update an existing Managed HSM. Only Tags can be updated currently.
+        /// Update an existing Managed HSM.
         /// </summary>
         /// <param name="existingManagedHsm">existing Managed HSM</param>
         /// <param name="parameters">HSM update parameters</param>
@@ -558,9 +596,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
 
             //Update the vault properties in the object received from server
             var properties = existingManagedHsm.OriginalManagedHsm.Properties;
-
-            // None property is allowed to be updated currently,
-            // Can be added here in the furture
+            properties.EnablePurgeProtection = parameters.EnablePurgeProtection;
 
             var response = KeyVaultManagementClient.ManagedHsms.Update(
                 resourceGroupName: existingManagedHsm.ResourceGroupName,
@@ -595,13 +631,11 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             {
                 KeyVaultManagementClient.ManagedHsms.Delete(resourceGroupName, managedHsm);
             }
-            catch (CloudException ce)
+            catch (ManagedHsmErrorException ce) when (ce.IsNotFoundException())
             {
                 // there's a known issue that the long running delete operation will
                 // finally throws an not found exception,
                 // we'll just ignore it
-                if (ce.Response.StatusCode != HttpStatusCode.NotFound)
-                    throw;
             }
         }
 
@@ -619,13 +653,11 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
 
             try
             {
-                KeyVaultManagementClient.Vaults.PurgeDeleted(managedHsmName, location);
+                KeyVaultManagementClient.ManagedHsms.PurgeDeleted(managedHsmName, location);
             }
-            catch (CloudException ce)
+            catch (ManagedHsmErrorException ce) when (ce.IsNotFoundException() || ce.IsNoContentException())
             {
-                if (ce.Response.StatusCode == HttpStatusCode.NoContent || ce.Response.StatusCode == HttpStatusCode.NotFound)
-                    throw new ArgumentException(string.Format(PSKeyVaultProperties.Resources.DeletedVaultNotFound, managedHsmName, location));
-                throw;
+                throw new ArgumentException(string.Format(PSKeyVaultProperties.Resources.DeletedManagedHsmNotFound, managedHsmName, location));
             }
         }